Update BillItem model to change quantity type from float to integer; add SoftDeletes trait to StockMovementLayer model; refactor InventoryCostingMethodSwitchService to optimize layer consolidation with database transactions; enhance InventoryDocumentReversalService to reverse layers based on specific GRN item IDs; adjust LandedCostService to use enum value for status; modify OpeningStockEntryService to lock stock records during updates; improve PaymentService to account for order discounts in bill calculations; and update BadDebtService to generate unique voucher numbers for write-offs.

+++ app/Models/BillItem.php
@@ -35,7 +35,7 @@ class BillItem extends Model
         'warehouse_id' => 'integer',
         'unit_id' => 'integer',
         'tax_id' => 'integer',
-        'quantity' => 'float',
+        'quantity' => 'integer',
         'rate' => 'float',
         'tax_amount' => 'float',
         'discount_amount' => 'float',
